+++ rust/sedona-spatial-join/src/lib.rs
@@ -18,7 +18,7 @@
 pub mod evaluated_batch;
 pub mod exec;
 mod index;
-mod join_evaluator;
+mod join_provider;
 pub mod operand_evaluator;
 pub mod partitioning;
 pub mod planner;
